# Conflicts:
#	features/merchant/seller_home_common/src/main/java/com/tokopedia/sellerhomecommon/presentation/view/viewholder/TableColumnHtmlViewHolder.kt

+++ features/logistic/logisticcart/src/main/java/com/tokopedia/logisticcart/shipping/model/ShipmentCartData.java
@@ -28,7 +28,7 @@ public class ShipmentCartData implements Parcelable {
     private String ut;
     private int insurance;
     private int productInsurance;
-    private int orderValue;
+    private long orderValue;
     private String categoryIds;
     private int preOrderDuration;
     private boolean isFulfillment;
@@ -180,11 +180,11 @@ public class ShipmentCartData implements Parcelable {
         this.productInsurance = productInsurance;
     }
 
-    public int getOrderValue() {
+    public long getOrderValue() {
         return orderValue;
     }
 
-    public void setOrderValue(int orderValue) {
+    public void setOrderValue(long orderValue) {
         this.orderValue = orderValue;
     }
 
@@ -221,7 +221,7 @@ public class ShipmentCartData implements Parcelable {
         dest.writeString(this.ut);
         dest.writeInt(this.insurance);
         dest.writeInt(this.productInsurance);
-        dest.writeInt(this.orderValue);
+        dest.writeLong(this.orderValue);
         dest.writeString(this.categoryIds);
         dest.writeInt(this.preOrderDuration);
         dest.writeByte((byte) (isFulfillment? 1 : 0));
@@ -246,7 +246,7 @@ public class ShipmentCartData implements Parcelable {
         this.ut = in.readString();
         this.insurance = in.readInt();
         this.productInsurance = in.readInt();
-        this.orderValue = in.readInt();
+        this.orderValue = in.readLong();
         this.categoryIds = in.readString();
         this.preOrderDuration = in.readInt();
         this.isFulfillment = in.readByte() != 0;
