+++ features/merchant/review/src/main/java/com/tokopedia/review/feature/historydetails/presentation/viewmodel/ReviewDetailViewModel.kt
@@ -20,7 +20,7 @@ class ReviewDetailViewModel @Inject constructor(private val coroutineDispatcherP
                                                 private val inboxReviewInsertReputationUseCase: InboxReviewInsertReputationUseCase)
     : BaseViewModel(coroutineDispatcherProvider.io) {
 
-    private val _feedbackId = MutableLiveData<Int>()
+    private val _feedbackId = MutableLiveData<Long>()
 
     private val _reviewDetails = MediatorLiveData<ReviewViewState<ProductrevGetReviewDetail>>()
 
@@ -29,8 +29,8 @@ class ReviewDetailViewModel @Inject constructor(private val coroutineDispatcherP
     val reviewDetails: LiveData<ReviewViewState<ProductrevGetReviewDetail>>
         get() = _reviewDetails
 
-    val feedbackId: Int
-        get() = _feedbackId.value ?: 0
+    val feedbackId: Long
+        get() = _feedbackId.value ?: 0L
 
     val submitReputationResult: LiveData<ReviewViewState<Int>>
         get() = _submitReputationResult
@@ -41,7 +41,7 @@ class ReviewDetailViewModel @Inject constructor(private val coroutineDispatcherP
         }
     }
 
-    fun getReviewDetails(feedbackId: Int, isRefresh: Boolean = false) {
+    fun getReviewDetails(feedbackId: Long, isRefresh: Boolean = false) {
         if(isRefresh) {
             _reviewDetails.value = LoadingView()
         }
@@ -60,7 +60,7 @@ class ReviewDetailViewModel @Inject constructor(private val coroutineDispatcherP
         _feedbackId.notifyObserver()
     }
 
-    fun setFeedbackId(feedbackId: Int) {
+    fun setFeedbackId(feedbackId: Long) {
         this._feedbackId.value = feedbackId
     }
 
@@ -68,7 +68,7 @@ class ReviewDetailViewModel @Inject constructor(private val coroutineDispatcherP
         return userSession.userId
     }
 
-    fun submitReputation(reputationId: Int, reputationScore: Int) {
+    fun submitReputation(reputationId: Long, reputationScore: Int) {
         _submitReputationResult.value = LoadingView()
         launchCatchError(block = {
             val response = withContext(coroutineDispatcherProvider.io) {
