+++ features/recharge/digitalproduct/src/main/java/com/tokopedia/digital/newcart/data/entity/requestbody/atc/Attributes.kt
@@ -19,7 +19,7 @@ class Attributes {
     var productId: Int = 0
     @SerializedName("order_id")
     @Expose
-    var orderId: Int? = null
+    var orderId: Long? = null
     @SerializedName("device_id")
     @Expose
     var deviceId: Int = 0
