+++ features/transaction/atc_common/src/main/java/com/tokopedia/atc_common/domain/model/request/AddToCartMultiParam.kt
@@ -5,15 +5,15 @@ import com.google.gson.annotations.SerializedName
 /**
  * Created by fwidjaja on 04/11/20.
  */
-data class AddToCartMultiParam (
+data class AddToCartMultiParam(
         @SerializedName("product_id")
-        val productId: Int = -1,
+        val productId: Long = -1,
 
         @SerializedName("product_name")
         val productName: String = "",
 
         @SerializedName("product_price")
-        val productPrice: Int = -1,
+        val productPrice: Long = -1,
 
         @SerializedName("quantity")
         val qty: Int = -1,
@@ -31,4 +31,5 @@ data class AddToCartMultiParam (
         val warehouseId: Int = -1,
 
         @SerializedName("category")
-        val category: String = "")
+        val category: String = ""
+)
